Move VAT tests to own file
